Statically initialize as many Parsing patterns as possible
Since this class gets re-initialized every time a document is parsed, this seems like a nearly-free perf improvement for any uses that do more than one document per process.

Many of the remainders appear to only have two variants, based on intellijDummyIdentifier - perhaps those would be worth duplicating explicitly and also statically initializing?  If the extra up-front cost is a concern, the two variants (and even the ones in this commit) could be moved to a lazy static initializer, so most(?) uses would only pay for the single pattern-set they use.